requirements.txt
Pillow==9.5.0
Pillow==10.1.0
streamlit==1.28.0
